+++ src/nanodsp/__init__.py
@@ -1,3 +1,3 @@
 """nanodsp - Python bindings of a set of dsp libraries via nanobind."""
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"